Remove deploy_test.sh
This basically runs 3 commands and can be a simple shell one-liner.

More importantly though, I’ll commit a change in a minute that will
allow both of the other scripts (flash-firmware.sh and sync-to-board.sh)
to use a EWHOME_PORT environment variable or a command line parameter,
including error handling and usage information, and it would just be
overkill to insert that into this script as well.